requirements.txt
numpy==1.11.2
matplotlib==1.5.3
nose==1.3.7
Pillow==3.4.2
